Movement framework is ready
Finished up a system for controlling movement, both among the player and other monsters. Input is now stored in a queue, and can be extracted by the system at a later date. Movement works, and correctly identifies if a tile is a wall or not. Logging has been partically set up, but is mostly just stubs. Enums all got shifted to one place for better access and readability. New enum for player actions (just movement right now), and new parts to the Direction enum (diagonals). Tiles got a little better about knowing when to be on and off. LOS got a lot smarter thanks to the player loop, and we're back to ~150 fps!