Started pathfinding! ~70% of the way to A*
Added in a new FastPriorityQueue implementation by BlueRaja. Needs to be accreditted before this project is associated with anything, for now fair use covers us. System can sucessfully use the queue to find and push new nodes to the search graph, and the take lowest priority nodes out of the queue to search. New spaces have been add, counting for Euclidean, Chebyshevian, and Manhattan types of movement. Spaces need to be standardized to just come from the Map structure, and the pathfinding needs an upgrade so that it knows when to stop and return a Path object. Right now that system has no heuristic, which will be changed in the next push.