Inventory added!
Added in a new component, the inventory. Holds items in stacks, manages them, and passes them along. Monsters and players that need to have items use this!

Also added a large number of ui panels, and new functionality to go with them. UI's can now layer on top of each other, and be exited with escape. InputTracking now records all input, so that item letters work. Item letters now work internally, so items that are picked up are assigned the first available letter. Items on the floor do not follow this rule. New component was added to track item visibility on the floor, based off of events added into the inventory. These events aren't hooked up into  the status effect system yet, but they will be soon.