Introduction of the "Status Effect" class
Created a new class named "Status Effect". These are essentially empty classes that carry an Effect, and some (very small) metadata about it's type. The big change is the PropertyDrawer that comes with them: It can automatically identify their type, identify available types of effect, switch effect types with a button click, handle background creation / deletion of the ScriptableObjects, and most importantly render the scriptable object editor as a dropdown to the variable. Now, effects can pretend to be crazy modular on the frontend and gain all the user benefits that follow that, while still being as powerful and functional on the backend as they were before. The only downside is that it's now slightly harder to track which effect belongs where in the backend, and list-copying-wonkiness can leave some effects hanging. Just pretend that they're regular variables, though, and it works fine.