Merged improvements, added more actions
Brought over many of the equipment improvements that Lab 8 made, as well as some of the nice behavioral changes and the pathfinding fixes. Added the EquipAction and RemoveAction, which both work in some capacity but are not fully integrated yet. Equipping as a concept needs to be reworked in order for it calling other actions (like removal) to make sense. Additionally, some functionality will need to be added to actions in order to handle things like exit / failure states that can propagate up the chain of actions.