Added Status Effect List UI
Status effects needed special care in order to be properly created and destroyed, the main problem being that they do not duplicate well and they do not get deleted well, due to their lingering references. This update adds a new data type, the StatusEffectList, which is literally just a List<StatusEffect> with some fancy new property drawer tech. This class handles all of the creation and deletion of new items, allowing status effects to once again work as intended, even in lists.