Status Effect Menu Dropdown
Status effects now use a Generic Menu, tied with a custom attribute, in order to display the available options as a nested folder structure when selected in the editor. This is sick. This should be the last of the massive changes for status effects, meaning this section of the redesign is done! The overall status effect workflow is much cleaner and easier to work with now. Expect slight improvements to behavior in the future, but nothing as extensive as the last few updates.