Added GameAction state check yield
Added a new option in GameAction, which is GameAction.StateCheck. When this object is yielded by any action, it will let the main game loop perform a check to determine if the turn should be ended, without skipping a frame. This allows you to guarantee that an action should be allowed to run after you subtract energy, without losing frames anymore!