Custom code wizard for Resources
Moved resources to their own file, Resources.cs. Added in a code wizard that generates this file from a template, using all the enums defined under Resource to create the code. This makes things nice and clean to use, because we can guarantee that core functionality (like health and mana tracking) is always around, but everything can still be accessed using ResourceList[Resource.ThingYouWant]. This will be great when abilities get done!

Also in here are some modifications in here to make it resistant to changes to the enum, at the cost of some readability. Sorry! The nicer code has been left in there, but since no one should be reading this file anyways, I left the more usable code instead.