First step in status effect rework
Connections for status effects are now pulled out into their own class, which makes it easy to extend and attach to more than just monsters. This class contains all of the connections, as well as a new "BlendInvoke" function that blends the invocation of two connection classes together. Status effects now hookup using a pregenerated list of connections. This is done automatically with a new tool, which makes it easy to add new connections and change existing connection types. This isn't entirely done yet, but the setup and hookup sections are ready. This also allowed around a 6x speedup in connection times, and amortizes the cost of using reflection for this task. Yay for magic backend code!