Finished up the effect rewrite
The effects script template is now generated automatically as well. Everything is done for the rewrite! In total, status effects are now fully generated from the template scriptable object. All names, types, and descriptions are controlled, and can be modified quickly and easily. Connections are also set up automatically, so the only thing that will need to be updated are old status effects!