Stairs work correctly now
Stairs now correctly hook up to individual entrances and exits, rather than the same entrance and exit that they represent. This system now allows for 1 way gates, as well as jumps to different branches (and back!). Added in some extra functionality to the GameController for waiting for a certain level index to be loaded, in case the game wants to expend some of the time the player will expect for level generation to create more of a buffer.