Added Dijkstra and improved some of the level loading
Added in a new Dijkstra map search, which should be very useful for item and monster placements. No idea if it actually works. I've also added some quick checks to the preloading section of the game controller, which prevent the preload from happening if it would never finish or it would take unreasonably long. This seems to have introduced a bug, which a one-frame pause after level finishing fixes.