Item spawning is ready!
Items are now organized into Loot Tables, which contain sets of items and upgrades that could be broadly applied to any item in the set. Items also have personal sets of upgrades that they could possess. ItemRarity and depth were added to items, which control where items are allowed to spawn, and how frequently they do. Loot table are condensed into loot pools, which follow general rules for controlling the types and rarities of items that spawn. These pools also do their best to recover when types and rarities requested can't be granted, escalating rewards up or down to match as best they can. This system is mostly battle-ready, and should be more than flexible enough for a real game!