First pass at Query Language
First pass at the query object, which aims to give a monster some info on the world around it. Queries are built of query terms, which have subjects and a question about those subjects, as well as weight. When a query is evaluated, each term is evaluated, and then all terms are added together and divided by total weight. Queries will be fully integrated into monsters, abilities, and AI in the next update!