Second pass at AI integration
AI's now can fight! They sit around until the player is there, then do a blind rush towards them. Combat actions got slightly reworked, so they currently steal energy if they are performed when they shouldn't be. This yells loudly, and should be removed for the actual build (replaced by a check if monsters can attack before they do). Fixed a major bug with stair generation that caused stairs to rarely spawn on top of each other. Added in querying for spells, and a function that compiles the best spell choice from the set. Updated queries to use a custom weight sum, which gets truncated back to 1 if it's greater, which gives queries a lot more flexibility. Now, they can be high ranking if ANY condition happens, vs all conditions, and a variety in between. Fixed a bug with map setup happening on every level load, not just initial loads. This fix won't actually do anything until monsters spawn in the generation pipeline, but it will be nice then.