More Auto-Combat is ready
Monsters now know how to flee, although they are very prone to running themselves into corners, even when better options are available. The flee code will need to be improved to allow monsters to the farthest available point, which will likely require a new pathfinding mode that uses the djikstra map as a cost map for A*. Shouldn't be too bad, but it might get a little gnarly. Player now has auto-attacking features, and will target and melee the nearest enemy. The move to nearest stair feature now works for both types of stairs, as well as only working for known stairs. This is possible due to a new pathfinding feature that uses Djikstra maps to search for the nearest goal from any location. This currently builds the full map, even if it doesn't need to, which is an area for further optimization down the line. Works great for now, though! The stair movement actions are also reworked on the player side, to require a double input before activating. This should be paired with a highlight and console message, to make things clear and prevent accidental pathfinding. This whole thing should also be reworked into a custom action that relies on PathfindingAction, which would more concisely control the player state and also do a nice job of preventing auto-move deaths. LOS updates for players and monsters were also improved, with LOS now re-collecting monsters every turn instead of fully updating like it used to. This is definitely the way to go in terms of efficiently allowing all monsters to know their field of view. Death and monster item drops were also added. Maps now know how many stairs they have, which is useful for pathfinding to them.