Fleeing works correctly now!
Fleeing now uses this slick Dijkstra map trick to better calculate flee maps. The new maps should help to prevent fleeing monsters from being stuck in the back of rooms, as well as improve the behavior of fleeing from multiple enemies. There is definitely room to improve this feature, but it works well enough for now!