Fixed the seed bug
Fixed a longstanding and slippery bug that was giving the appearance that the dungeon generator seeding incorrectly on startup. This turned out to be an issue with instancing scriptable objects, which caused an object to incorrectly save it's size between runs. This would cause some shared machines to grow too large, and incorrectly save that size when being used for a smaller machine. This would cause a smaller machine to be able to move itself out of bounds on the first run, but only the first run. Machines now resize themselves before their size is necessary, as well as instancing themselves to prevent this type of problem in the future.