Monsters can cast spells
Monsters can now cast spells! It's definitely a little messy still, and there is some work to be done on cleaning the code and doing better target selection for spells with both range and area. With that said, the current system works more than well enough to see some intelligence in the monsters. Monsters now find the highest priority spell, and will cast it if it seems like the logical thing to do. They'll spam spells until they run out of mana, at which point they currently just run at the player and die. This system probably won't be upgraded till midway on the development of the next game, but it's stable enough to pull it's weight till then.