Interactable tiles are ready to go, monsters know how to use them
Interactable tiles are finished code-wise, and monsters can now intelligently select them and use them. Fleeing was improved a little bit to encourage prevent fleeing monsters from freezing in place forever once cornered. Added a demo healing station, which monsters know how to use to return to full health. Monster choices are now pushed through a priority queue - it might be worth it to put all choices in here, so that the numbers can be fudged a little at the end!