RexPaint prefabs are ready for floors
First pass at the RexPaint prefab - Added in the SadRex library, and did a lot of finagling to get everything up and running. Created a new ScriptedImporter that basically serializes a rexpaint file to a TextAsset, and made the reverse to read it at runtime. Discovered AssetBundles, but they weren't really useful so they got dropped. Added in a new button (windows only at the moment, sorry) that launches a rexpaint instance hooked up to the right directory. This ~should~ make working with these pipelines nice and easy, and make installation for windows a breeze as well. High on the priority chain is finding a way to do this for Mac.