Clean up some systems
Cleaned up the remains of the rexpaint launcher, so the RexPaint folder is now completely dedicated to the program and it's images. RexpaintAssetPipeline got the loading code moved into it, so saving and loading is all in the same file now. Pulled out some extraneous printing statements that were no longer displaying helpful information.