Monster spawning and loadouts are ready to go!
Added in a new loadout system that equips monsters with gear and abilities. Added in a new monster spawning system, which mimicks the item spawning system (with some difference in how cutoffs are generated). The item spawning system could probably use an overhaul - I like how this code works a lot better, and it doesn't seem to really be a problem on performance. Once rexpaint prefabs support monsters, this whole section will be ready to go!