First pass at animation system
Unity's animation system is, as far as I know, missing two major functions which we needed for this game. The first is queuing, where animations on different objects are allowed to queue up against each other, or take a spotlight position. The other is dynamic animations, where code can be driven like an animation is. This update adds a new Animation Controller, along with some associated animations, that do just that. Now, all monsters bounce to the beat as they move, and attack by gently bumping into their friends. The first of the spell animations has also been added in the form of the ExplosionAnimation. To facilitate the dynamic behavior of these animations, a new type has been added: the SpriteGrid. SpriteGrids are fast, light objects that allow for extremely large grids with quick updates. They control tiles by editing the UV's of the underlying mesh, and batching together many of those UV operations at once. This kicks the majority of the work to the GPU, keeping them super fast across large tasks. Even better, they're entirely procedural, and only require 1 gameobject in order to run thousands of sprites. The TargetingPanel has also been updated to use a SpriteGrid in the background, solving the lag problem that was previously occuring and bringing the render and update times down to under a tenth of a millisecond. Up next, all sorts of different procedural animations for anything and everything!