New targeting features + bugfixes
Targeting now correctly responds to hidden tiles - they won't be highlighted on the view or animate, but will correctly collect targets. This means that you can't use certain features, like LOS_AREA targeting, to scout terrain - but you can use the resulting explosion to tell if monsters are hiding around the corner, which is a great hidden feature that I want to preserve. Fixed the bug where items could appear on top of monsters. MoveAction now internally moves monsters as well, making it easier to track in the editor and making sure the first frame of their animations is consistent.