Resource Rebuild
Rebuilt the resource system from the ground up, using the same type of system as the UE Gameplay Ability system. The core idea is to not separate stats and resources into different components, but to store everything into one dictionary. All values are implicitly stored as zero if they are not in the dict, and otherwise they hold their value. This representation holds the same speed for item lookup, and drastically improves the speed of new value additions. Added a custom property drawer as well, so it's easy to add new properties. As a benefit, the property is only as big as the number of modified items, which makes things nice for modification-centric field like equipment and effects!

Next steps are repeating this process for weapon/spell modifications, which need an overhaul as well and would benefit from an almost identical set of upgrades. It isn't critical, but some more UI upgrades might be in the works - a system (using attributes) to tag the enums with a folder structure could be helpful if the number of stats ends up getting absurdly high.