Added RNG functions and unit-y testing
Added in a slew of new RNG functions that encapsulate a variety of useful distributions. Added in the unity testing suite, and created a tests for each of the new distributions. This stuff isn't super useful yet, but should end up being helpful with some of the new level generator redesigns. Also moved scripts and plugins into the 'framework' folder, to set up for distributing this framework to other games.