Move monster info into branches
Removed monster pools, and add the monster info into branches instead. Monster pools were originally supposed to be a composing feature, which had been superseded by the concept of branches and become an extra step assignment that wasn't needed. Relevant spawning info (like unique chances and OOD spawning) are in branches as well, which should help to make the per-branch feel more consistent and be a more powerful tool for shaping the feel of the dungeon.