Moved item info into branches + optimizations
All of the item spawning is now per-branch as well. This update got the intended behavior from a game and developer side, but is still pretty messy in the background. Some other update will be needed in the future to dive into that old mess and optimize it, but for now it all runs acceptably fast. Dungeon machines were moved into being coroutines, which allowed for better splitting of the work between frames. The levelLoader msPerFrame time allocation now uses the actual number, which works a lot better with the improved granularity. RexRoom glyph conversions now use a dictionary for lookups, which seems to have halved the generation time. SimpleRoomMachine was found to be allocating a ridiculous amount of space while running - this was mitigated by the other fixes but not solved, and will need to be drastically changed in the future.