More dungeon optimizations
Added in more optimizations to the generator. The room machine now uses all of it's attempts to place a single prefab, rather than one prefab per attempt. Identified other GC problems in the map building sequence - something is being allocated A LOT, and is causing mild GC spikes after that coroutine finishes. The Djikstra pathfinder (and by extension, probably the A* one as well) end up creating allocs for every space they enter. This is causing slowdown and lags spikes in generation, and is also a likely candidate for why the AI's are slower than expected.