Added priority queue
The previous priority queue implementation was not made in-house, and as such had weird behavior like having to allocate a class object every time you wanted to add a node. That made no sense for this project, so that was removed and a new PriorityQueue<T> type was added. This mimics the C# priority queue type, which we can't have because our .NET version is not up to date. The new system uses an array in background as a binary heap, and gives us O(log(n)) queue and dequeue and O(1) clear time. It also reuses memory from past allocations, allowing it to essentially cache that memory to prevent new allocs or garbage collects. This improvement was added to the pathfinding system and the monster AI, cutting down allocs per search from ~11K to ~50. I didn't measure the exact speedup, but a quick test showed improvements with ~20 actively in-combat monsters running very smoothly. The new system is a lot more flexible for upgrading in the future, which should make the eventual 3D pathfinding more painless.