Convert Machines + Overrides
Converted machines to be just regular objects instead of ScripableObjects. Added a new extension to the Juce editor to support that automatically. Set all of the branches to use the new system, and tested that things actually generate. Completed level overrides by adding in new types of changes that are available, and the fields to support them. Added in a custom property drawer that modifies which elements are shown, and extended that property drawer to visually show the exact elements that will be added or deleted. With the aside of the eventual branch lookdev that needs to happen and some minor convenience changes, branches are structurally complete, and the dungeon update is pretty much done!