Solved tech for cheat menu
Added in a new cheat menu, which gets compiled for release builds. The cheat menu uses reflection to find methods tagged with the [Cheat] attribute, and stores some info about what kind of string represents them. The player can input command strings into the cheat menu, which get parsed using a TypeConverter to transform the strings into the native type the function requests. Needs a lot of cleanup work on the UI bit, but it's very functional for now. Added in cheats to kill and heal the player, reveal the map, and remove all monsters from the map.