Added first pass of floating panels
This commit is fairly unstable and very WIP. Added in new class to implement and manuever floating panels. Floating panels are instantiated with a desired size, and a constraint solver attemtps to fit them onto the screen, minimizing overlap and distance to goals. As of right now, that solver is based on a simple heuristic, and uses a pretty basic process to spread them out. This works alright for small numbers of rects, but has a lot of room to improve in terms of speed and stability. The end goal for this system is to be able to safely ask for on-screen dialogue boxes, which can then be filled with all sorts of cool information.