Added in Quadtrees
Needed a quadtree implementation for the item spawning rebuild. Created a generic quadtree class to support two dimensional lookups using Rects. Added in new tests to the Suite to confirm that the trees work and are fast. The trees perform consistently with what I would expect from a quadtree. When testing a standard use case (~150 items in a 100x100 zone), the tree was able to retrieve items at particular zones in about .002 ms, which is way faster than I need it to be. One thing the trees don't support is moving items. That will need to be built in at some point to support monster query information, but for now it's fine. Next update will be improvements to the item spawning system using this new tech.