Moved item spawning to quadtrees
Items are now grouped by branch, and use a quadtree for spawning rather than the weird table-chopping that existed before. Dungeon gen is about the same speed, but the changes should pay for themselves once there are more items to choose from in the game. In the short term, branches should be able to have a lot more item variety internally without needing to worry about losing speed from extraneous items, and queries of the form "Give me an item from this branch" are now possible. Aside from making some convenience functions down the line, the dungeon spawning system are good to go! Next update will hopefully be a medium-sized overhaul to the cost logic of pathfinding (again), to support dynamic tiles and multi-tile monsters.