Added Dungeon Systems
Added in dungeon systems, which are a new interface for objects that want to live above levels, branches, and the world to control general game behavior. Also added in a property drawer for a drop-down selector, and tested them out to make sure they worked correctly. The main idea with these is ot make it easier to add more level-specific gimmicks that can't be efficiently implemented with just tiles - think effects like the hell effects and abyss effects in crawl. These systems get regular updates about game state and are hooked into to all the major systems, and activate and deactivate automatically as they move in and out of scope.