Set up activatable items
Moved usable items from TargetableItem and ApplyableItem to ActivatableItem. This new class functions as a stand-in that ties an ability to an item, and can optionally consume the item on ability use. This does the exact same job that the old classes did, but with a wider breadth of use cases. The current setup is somewhat annoying, as ability instances need to be created for every single item. That will definitely be fixed in the future, with some sort of middleman class or a selector implementation.