Finished clockwork dungeon generator
Added in a few graph tools to make finding good edge connections easier. Clockwork dungeons now generate with a ring of conveyor belts created by gift-wrapping, and clean up the stragglers with an minimum spanning tree that gets the center of the graph connected. This has very nice loops and some interesting gameplay consequences, so I'm pretty happy with how it's turned out.