Added in Nightmare BSP
Named for both the branch and the implementation. New dungeon generator variant that does BSP to carve the map into hallways and big habitable zones, then turns those zones into apartment-like clusters of rooms and hallways. Turned out WAY better than expected, I'm really happy about this one. There's lots of rooms for efficiency gains, but that's a future me problem. All that's left for the nightmare branch is the visual pass and shaders, along with setting up the abilities and passives of it's inhabitants. Almost ready!