Added Husk Node
Added new type of monster to nightmare that can scream to call in husks. These husks get spawned in the same way that the nightmare does, so it should hopefully reinforce the idea of exploring with an intention of creating safety zones. Each node can spawn up to 8 husks, with husks showing up 5 rounds after a scream. This new ability is also caller-agnostic, so if/when players can steal a monster ability they'll be able to use this mechanic as well. (Confirmed functional through debug mode)