Fixed LOS bugs
LOS had some super weird edge cases where it wouldn't correctly retrieve monsters that were in sight. The root cause of this ended up being that LOS wasn't updating frequently enough, as some of the preconditions of the system where it was not recalculated every turn were not actually true. This led to instances where things like doors could have asymmetric sight, because one monster sees an LOS where the door was closed (and therefore more visible against a wall segment), while the other sees it as open (and therefore can't see the first monster while standing in it). The blanket solution to these problems is to recalculate LOS at the start of every local turn. This has completely wiped all of the bugs, and doesn't seem to have impacted performance. There are many more effeciency gains that could be done, but for now I'm not touching it until it causes an FPS problem.