Added localization
Added in localization package that does a bunch of crazy stuff. Effects now have dynamic string names and descriptions, which use reflection to generate their values. Added english locale, and set it as default. Localized all current monster and players, as well as all of the effects available from items. Next steps are showing monster strings in some form of inspector, and adding in the rest of the effects to the system.