Weapon brands
Reworked the background tech for weapon brands, and came up with a lot of new solutions for expediting the process in the future. Added lots of new weapon types, with special forms of attack and movement. Added legendary items to confirm that the spawn system will actually make them. Fixed a variety of item effect bugs that were previously not seen because items had no effects.