Rework for overheating
Changed how the overheat mechanic works to be less annoying. The punishment damage is about the same, but stops much earlier and drains player heat as well. This is very open to change with more testing, but should prevent players from getting stuck overheating and instead make the mechanic a punishing melee reset.