Fix for infinite stats and HM Tempo
Fixed some internals for how stats are added. Stats now use the &= and ^= operators for addition and subtraction in place, rather than += and -=. This frees up the plus operator to function as expected, creating a new variable that can be assigned. The new operators allow us to keep the same performance benefits that orginally caused the change, and hopefully prevent this bug from occuring again in the future. Also fixed a bug where the heat mage timer would reset on self-casts.