Add LOS improvements + new target type
Switched LOS to storing enemies and friendlies in different arrays. This was the most common use case by far, so it made much more sense to just do that every time (constant time cost) vs recalculating it every time a query needed it. Used the new tech to build a new targeting type which uses all monsters in sight as the base points for area generation.