Added machines to world gen + blackboard
Added in a set of machines that are allowed to run at world gen. Added in a static world object that persists throughout the game. Added a new blackboard abstraction to the world, which allows for storing arbitrary objects under string names. Added a machine that generates a jungle path, and an effect that can read it off. Created the first codex object to apply that effect. Fixed a small bug where monsters could stall out too early.