Add in city of gold and locking mechanism
Added in a new branch system that records the player's traversal order to determine if they should be allowed in the city of gold. Used the locking system to enforce that rule, and prevent the player from entering without the ordering. Not sure if I love that you can see the exit right now, but that can come later.