Jungle Monsters + Flaming + Spawn Item
Added first 4 jungle monsters, but without any of their set dressing. Added in the hydranha, which has a version of the really great hydra effect from DCSS. Added flaming brand to interact with that effect. Added in the SpawnItem cheat, and did a lot of work to get that functioning nicely.