First pass at saving with Odin
Pulled in the open source Odin Serializer, and used it create a basic save system. Added a new data arena class that stores linear data, and can index it with a handle. The handles are very small and serialize well, so the whole structure allows for easy control of objects that may have multiple reference, self references, and circular references. Odin can handle those, but only when they are all serialized at the same time, which leaves room for error that I'd like to avoid. Next steps are moving references over to the handle system for initial testing.